Able to now call both models, with error checking, timeout, and retries API calls a certain number of times before stopping
