+++ package.json
@@ -29,7 +29,8 @@
     "react-hook-form": "^7.47.0",
     "tailwind-merge": "^1.14.0",
     "tailwindcss-animate": "^1.0.7",
-    "zod": "^3.22.4"
+    "zod": "^3.22.4", 
+    "@types/node": "^7.0.5"
   },
   "devDependencies": {
     "@types/node": "^20.7.0",
@@ -46,6 +47,7 @@
     "postcss": "^8.4.30",
     "tailwindcss": "^3.3.3",
     "typescript": "^5.2.2",
-    "vite": "^4.4.9"
+    "vite": "^4.4.9",
+    "node-fetch": "^3.3.2"
   }
 }
